# MNT Refactors PredictorTree to use vector of node_structs