--- a/sklearn/ensemble/_hist_gradient_boosting/_grower.pyx
+++ b/sklearn/ensemble/_hist_gradient_boosting/_grower.pyx
@@ -2,6 +2,7 @@
 # cython: boundscheck=False
 # cython: wraparound=False
 # cython: language_level=3
+
 import numpy as np
 cimport numpy as np