--- a/sklearn/ensemble/_hist_gradient_boosting/_predictor.pxd
+++ b/sklearn/ensemble/_hist_gradient_boosting/_predictor.pxd
@@ -7,7 +7,7 @@
 from .common cimport node_struct
 from libcpp.vector cimport vector
 
-cdef class PredictorNodes:
+cdef class TreePredictor:
     cdef vector[node_struct] _nodes
     cdef node_struct* get(self, int node_idx) nogil
     cdef int get_size(self) nogil