--- a/sklearn/ensemble/_hist_gradient_boosting/_predictor.pxd
+++ b/sklearn/ensemble/_hist_gradient_boosting/_predictor.pxd
@@ -8,6 +8,6 @@ from .common cimport node_struct
 from libcpp.vector cimport vector
 
 cdef class TreePredictor:
-    cdef vector[node_struct] _nodes
+    cdef public vector[node_struct] nodes
     cdef node_struct* get(self, int node_idx) nogil
-    cdef int get_size(self) nogil
+    cdef int get_n_nodes(self) nogil