--- a/sklearn/ensemble/_hist_gradient_boosting/common.pxd
+++ b/sklearn/ensemble/_hist_gradient_boosting/common.pxd
@@ -19,8 +19,6 @@ cdef packed struct hist_struct:
 
 
 cdef packed struct node_struct:
-    # Equivalent struct to PREDICTOR_RECORD_DTYPE to use in memory views. It
-    # needs to be packed since by default numpy dtypes aren't aligned
     Y_DTYPE_C value
     unsigned int count
     unsigned int feature_idx