Towards https://github.com/scikit-learn/scikit-learn/pull/16909

This PR refactors `PredictorTree` to use a vector of structs and not a numpy array to store the predictor nodes. This is done because categorical support needs to store a dynamic array in the predictor node. A secondary goal of this PR is to make  https://github.com/scikit-learn/scikit-learn/pull/16909 easier to review.

Running the higgs boson benchmark, this does not change performance:

### master

```
Fit 100 trees in 30.473 s, (3100 total leaves)
Time spent computing histograms: 15.321s
Time spent finding best splits:  0.103s
Time spent applying splits:      2.345s
Time spent predicting:           1.466s
fitted in 30.480s
predicted in 2.774s, ROC AUC: 0.8228, ACC: 0.7415
```

### This PR

```
Fit 100 trees in 30.394 s, (3100 total leaves)
Time spent computing histograms: 15.341s
Time spent finding best splits:  0.102s
Time spent applying splits:      2.372s
Time spent predicting:           1.458s
fitted in 30.401s
predicted in 2.749s, ROC AUC: 0.8228, ACC: 0.7415
```

CC @NicolasHug