@thomasjpfan 2020-08-13
BTW I have been thinking about this comment: https://github.com/scikit-learn/scikit-learn/pull/16909#discussion_r469408699