@NicolasHug 2020-08-13
Yes I agree that staying in numpy world would be best, but from what you've tried it seems it isn't possible here?