@thomasjpfan 2020-08-13
The only way that I see it being possible to use numpy arrays was to map categorical features to the bins in predict. (This was the very first implementation I tried)