@NicolasHug 2020-09-04
@ogrisel we'd appreciate your review here as this will make the categorical PR much cleaner (and I'd love to get this done before I switch jobs in October ;) )



According to benchmarks this PR does not seem to introduce any regression. Here are averages over 10 experiments of the following (trees have an average depth of 26):

```py
X, y = make_classification(n_samples=10_000, n_features=10, random_state=0)
est = HistGradientBoostingClassifier(max_iter=200, max_leaf_nodes=256,
                                     min_samples_leaf=1, early_stopping=False)
```

This PR

```
         predict        fit
count  10.000000  10.000000
mean    0.043068   5.715113
std     0.000502   0.519689
min     0.042545   5.413697
25%     0.042872   5.507002
50%     0.042967   5.554982
75%     0.043209   5.588134
max     0.044308   7.154663
```

Master

```
         predict        fit
count  10.000000  10.000000
mean    0.054503   6.808190
std     0.025802   0.806578
min     0.044969   6.305942
25%     0.045176   6.397328
50%     0.045298   6.551962
75%     0.047053   6.786230
max     0.127566   9.019717
```

<details>

```py
from time import time

from sklearn.datasets import make_classification
from sklearn.experimental import enable_hist_gradient_boosting
from sklearn.ensemble import HistGradientBoostingClassifier
import numpy as np
import pandas as pd

X, y = make_classification(n_samples=10_000, n_features=10, random_state=0)
est = HistGradientBoostingClassifier(max_iter=200, max_leaf_nodes=256,
                                     min_samples_leaf=1, early_stopping=False)

fit_times = []
predict_times = []

n_exp = 10

for _ in range(n_exp):
    print(_)

    tic = time()
    est.fit(X, y)
    toc = time()
    fit_times.append(toc - tic)

    tic = time()
    est.predict(X)
    toc = time()
    predict_times.append(toc - tic)

print(f"average depth={np.mean([pred[0].get_max_depth() for pred in est._predictors])}")

df = pd.DataFrame({'predict': predict_times, 'fit': fit_times})
print(df.describe())
```

</details>