@thomasjpfan 2020-09-04
From those numbers, it looks like this PR made predicting and fitting faster?