@NicolasHug 2020-09-04
Yes, by a small margin (which might not be significant, especially considering your benchmarks above). I'm being very conservative by just mentioning the lack of regression ;)